refactor: 이벤트 리스너 설정 및 기능 모듈화
- setupEventListeners 함수 추가로 이벤트 리스너 설정을 중앙 집중화
